Fixed card displaying system
Now the player can use all the cards that he has no matter how many he accumulates because they are displayed by order in terms of what card is on top of which. Also fixed the display system of the cards on the board so that the attacking cards is always on the bottom and the defending card on top. This fixes #17 and fixes #24